+++ pulse-binding/CHANGELOG.md
@@ -50,6 +50,8 @@
  * Converted `def::{SinkFlagSet, SourceFlagSet}` to [`bitflags`] types, moving the
    `def::sink_flags::*` and `def::source_flags::*` to associated constants of them respectively,
    deprecating the old constant sets.
+ * Converted `direction::FlagSet` to a [`bitflags`] type, moving the `direction::flags::*` constants
+   to associated constants of it, deprecating the old constant set.
  * Added `MonotonicTs::checked_add_duration()` and `MonotonicTs::checked_sub_duration()` along with
    corresponding `Add[Assign]` and `Sub[Assign]` impls for adding a `Duration` to `MonotonicTs`.
    This complements the existing functionality for adding a `MicroSeconds`, without requiring